fix(eslint-errors): menu-vertical
