+++ packages/menu-vertical/src/menu-vertical-context.tsx
@@ -1,88 +1,11 @@
 import React from 'react';
-import { MenuVerticalSubmenuButtonProps } from './menu-vertical-submenu-button';
-import { MenuVerticalItemProps } from './menu-vertical-item';
-
-export interface MenuItemTypes {
-  submenuItem?: React.ReactElement<MenuVerticalSubmenuButtonProps>;
-  menuItems: React.ReactElement<MenuVerticalItemProps>[];
-}
-
-export type MenuIndex = number | string | null | undefined;
-
-interface Submenu extends MenuItemTypes {
-  parentMenuId: string;
-  parentLiMenuIndex: MenuIndex;
-  submenuOpen: boolean;
-  setSubmenuOpen: React.Dispatch<React.SetStateAction<boolean>>;
-}
-
-interface Menu {
-  [id: string]: Submenu;
-}
-
-interface UseMenuVerticalPropsStates {
-  rootId: string;
-  rootMenuId: string;
-  menu: Menu;
-  active: MenuIndex;
-  focused: MenuIndex;
-  activeMenuId: string;
-  current: MenuIndex;
-  currentMenuId: string;
-}
-
-interface UseMenuVerticalPropsFunctions {
-  setMenu: React.Dispatch<React.SetStateAction<Menu>>;
-  setActive: React.Dispatch<React.SetStateAction<UseMenuVerticalPropsStates['active']>>;
-  setActiveMenuId: React.Dispatch<React.SetStateAction<UseMenuVerticalPropsStates['activeMenuId']>>;
-  setCurrent: React.Dispatch<React.SetStateAction<UseMenuVerticalPropsStates['current']>>;
-  setFocused: React.Dispatch<React.SetStateAction<UseMenuVerticalPropsStates['focused']>>;
-  setCurrentMenuId: React.Dispatch<React.SetStateAction<UseMenuVerticalPropsStates['currentMenuId']>>;
-  next: () => void;
-  prev: () => void;
-  getAboveSubmenuIds: (menuId: string, menuIds?: string[]) => string[];
-  setSubmenus: (defaultOpen?: boolean, options?: { openMenuIds?: string[]; closeMenuIds?: string[] }) => void;
-  setCurrentActive: (menuIndex: UseMenuVerticalPropsStates['current']) => void;
-  setCurrentActiveFocus: (menuIndex: UseMenuVerticalPropsStates['focused']) => void;
-  setActiveFocus: (menuIndex: UseMenuVerticalPropsStates['focused']) => void;
-}
-
-export type UseMenuVerticalProps = UseMenuVerticalPropsStates & UseMenuVerticalPropsFunctions;
-
-const defaults = {
-  rootId: '',
-  rootMenuId: '',
-
-  menu: {},
-  setMenu: () => ({}),
-
-  active: '',
-  setActive: () => ({}),
-
-  focused: '',
-  setFocused: () => ({}),
-
-  current: '',
-  setCurrent: () => ({}),
-
-  activeMenuId: '',
-  setActiveMenuId: () => ({}),
-
-  currentMenuId: '',
-  setCurrentMenuId: () => ({}),
-
-  next: () => ({}),
-  prev: () => ({}),
-  getAboveSubmenuIds: () => [],
-  setCurrentActive: () => ({}),
-  setCurrentActiveFocus: () => ({}),
-  setActiveFocus: () => ({}),
-  setSubmenus: () => ({}),
-};
-
-const MenuVerticalContext = React.createContext<UseMenuVerticalProps>(defaults); // context defaults are generated on render
-
-export const useMenuVertical = () => React.useContext(MenuVerticalContext);
+import {
+  MenuVerticalContext,
+  UseMenuVerticalProps,
+  UseMenuVerticalPropsFunctions,
+  UseMenuVerticalPropsStates,
+} from './context';
+import { MenuIndex } from './types';
 
 export interface MenuVerticalProviderProps {
   children: JSX.Element | ((parameters: UseMenuVerticalProps) => JSX.Element);
