+++ packages/menu-vertical/src/use-menu-vertical.ts
@@ -1,0 +1,4 @@
+import React from 'react';
+import { MenuVerticalContext } from './context';
+
+export const useMenuVertical = () => React.useContext(MenuVerticalContext);
